+++ hammer-sqldb/src/test/java/cn/featherfly/hammer/sqldb/pt/BenchmarkTestBase.java
@@ -20,7 +20,10 @@ import cn.featherfly.common.db.SqlFile;
 import cn.featherfly.common.db.dialect.Dialect;
 import cn.featherfly.common.db.dialect.Dialects;
 import cn.featherfly.common.db.dialect.PostgreSQLDialect;
+import cn.featherfly.common.db.metadata.DatabaseMetadata;
+import cn.featherfly.common.db.metadata.DatabaseMetadataManager;
 import cn.featherfly.common.lang.ClassLoaderUtils;
+import cn.featherfly.common.lang.SystemPropertyUtils;
 import cn.featherfly.common.lang.UriUtils;
 
 /**
@@ -34,30 +37,40 @@ public abstract class BenchmarkTestBase {
 
     protected Dialect dialect;
 
+    protected DatabaseMetadata metadata;
+
     @BeforeSuite
-    @Parameters({ "dataBase" })
-    public void init(@Optional("mysql") String dataBase) throws IOException {
+    @Parameters({ "dataBase", "url" })
+    public void init(@Optional("mysql") String dataBase,
+        @Optional("jdbc:mysql://127.0.0.1:3306/hammer_jdbc?characterEncoding=utf8&useUnicode=true&useSSL=false") String url)
+        throws IOException {
         DOMConfigurator.configure(ClassLoaderUtils.getResource("log4j_none.xml", this.getClass()));
 
-        initDataBase(dataBase);
+        initDataBase(dataBase, url);
+
+        metadata = DatabaseMetadataManager.getDefaultManager().reCreate(dataSource);
+
+        //        System.err.println(url);
+        System.err.println("jdk " + SystemPropertyUtils.getJavaVersion());
+        System.err.println("supportsBatchUpdates = " + metadata.getFeatures().supportsBatchUpdates());
     }
 
-    public void initDataBase(String dataBase) throws IOException {
+    public void initDataBase(String dataBase, String url) throws IOException {
         System.err.println("***********************************************");
         System.err.println("***********************************************");
         //        configFile = String.format(CONFIG_FILE_PATTERN, dataBase);
         switch (dataBase) {
             case "mysql":
-                initMysql();
+                initMysql(url);
                 break;
             case "postgresql":
-                initPostgresql();
+                initPostgresql(url);
                 break;
             case "sqlite":
-                initSQLite();
+                initSQLite(url);
                 break;
             default:
-                initMysql();
+                initMysql(url);
                 //                configFile = String.format(CONFIG_FILE_PATTERN, "mysql");
                 break;
         }
@@ -66,19 +79,14 @@ public abstract class BenchmarkTestBase {
     }
 
     //    @BeforeSuite(groups = "mysql", dependsOnMethods = "init")
-    public void initMysql() throws IOException {
+    public void initMysql(String url) throws IOException {
         System.err.println("initMysql");
         //        ConstantConfigurator.config();
         //        ConstantConfigurator.config("constant.mysql.yaml");
 
         BasicDataSource ds = new BasicDataSource();
-        //        dataSource.setUrl("jdbc:mysql://127.0.0.1:3306/hammer_jdbc?useUnicode=true&characterEncoding=UTF-8");
-        //        ds.setUrl(
-        //                "jdbc:mysql://127.0.0.1:3306/hammer_jdbc?serverTimezone=UTC&characterEncoding=utf8&useUnicode=true&useSSL=false");
-        //        ds.setUrl(
-        //                "jdbc:mysql://127.0.0.1:3306/hammer_jdbc?serverTimezone=CTT&characterEncoding=utf8&useUnicode=true&useSSL=false");
-        // 高版本mysql-connector已经不需要serverTimezone=CTT
-        ds.setUrl("jdbc:mysql://127.0.0.1:3306/hammer_jdbc?characterEncoding=utf8&useUnicode=true&useSSL=false");
+        //        ds.setUrl("jdbc:mysql://127.0.0.1:3306/hammer_jdbc?characterEncoding=utf8&useUnicode=true&useSSL=false");
+        ds.setUrl(url);
         ds.setDriverClassName("com.mysql.cj.jdbc.Driver");
         ds.setUsername("root");
         ds.setPassword("123456");
@@ -92,13 +100,14 @@ public abstract class BenchmarkTestBase {
     }
 
     //    @BeforeSuite(groups = "postgresql", dependsOnMethods = "init")
-    public void initPostgresql() throws IOException {
+    public void initPostgresql(String url) throws IOException {
         System.err.println("initPostgresql");
         //        ConstantConfigurator.config("constant.postgresql.yaml");
 
         BasicDataSource ds = new BasicDataSource();
         //        ds.setUrl("jdbc:postgresql://localhost:5432/hammer_jdbc");
-        ds.setUrl("jdbc:postgresql://::1:5432/hammer_jdbc"); // install postgresql in wsl with docker
+        //        ds.setUrl("jdbc:postgresql://::1:5432/hammer_jdbc"); // install postgresql in wsl with docker
+        ds.setUrl(url); // install postgresql in wsl with docker
         ds.setDriverClassName("org.postgresql.Driver");
         ds.setUsername("postgres");
         ds.setPassword("123456");
@@ -120,12 +129,12 @@ public abstract class BenchmarkTestBase {
     }
 
     //    @BeforeSuite(groups = "sqlite", dependsOnMethods = "init")
-    public void initSQLite() throws IOException {
+    public void initSQLite(String url) throws IOException {
         System.err.println("initSQLite");
         //        ConstantConfigurator.config("constant.sqlite.yaml");
 
         String path = new File(UriUtils.linkUri(this.getClass().getResource("/").getFile(), "hammer.sqlite3.db"))
-                .getPath();
+            .getPath();
         System.out.println(path);
         BasicDataSource ds = new BasicDataSource();
         ds.setDriverClassName("org.sqlite.JDBC");